requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13

# Auth
python-jose[cryptography]>=3.3
cryptography>=42.0
httpx>=0.27

# Config
pydantic-settings>=2.0
python-dotenv>=1.0

# CF interaction
playwright>=1.40.0
requests>=2.31.0

# LLM
anthropic>=0.40.0
openai>=1.0.0

# Misc
pyyaml>=6.0
